Upgrade Go + Dependency versions to latest
Release notes here: https://github.com/hashicorp/terraform-plugin-framework/releases

This is mostly a shuffling of package names for the terraform-plugin-framework.

I've renamed the provider struct to vercelProvider just so the name doesn't clash with the new terraform-plugin-framework/provider package.

I've also bumped the minimum go version to 1.19, as the latest tf-plugin is >1.18. I've had to disable the `tfproviderlint` tool to do this. It looks poorly maintained and still doesn't support Go1.18+ syntax.