Prevent project field coercion on creation
In the Vercel API the following fields are coerced to null during project creation
if they are falsy:
* BuildCommand
* DevCommand
* InstallCommand
* OutputDirectory
* PublicSource

This causes an issue when they are specified, but falsy, as the
terraform configuration explicitly sets a value for them, but the Vercel
API returns a different value. This causes an inconsistent plan error.

We avoid this issue by choosing to use values from the terraform state,
but only if they are _explicitly stated_ *and* they are _falsy_ values
*and* the response value was Null. This is important as drift detection
would fail to work if the value was always selected.