Omit `null` environment variables in vercel_deployment
Creating a vercel_deployment and passing a `null` environment variable
results in a conversion error.
```
An unexpected error was encountered trying to build a value. This is always
an error in the provider. Please report the following to the provider
developer:

Received null value, however the target type cannot handle null values. Use
the corresponding `types` package type, a pointer type or a custom type that
handles null values.

Path: ["BAZ"]
Target Type: string
Suggested `types` Type: basetypes.StringValue
Suggested Pointer Type: *string
```

We need to handle the case where the field value is `null`, which is
currently omitted.

Since the API only accepts string values for enviroment variables, the
best solution is to simply omit sending `null` to the API altogether.

Closes #95