Fix issues with removing git_repository from a project
At the moment, this should force recreation of the project, as we don't
support updating or removing the linked repository.

See issue #105 for info on this.

However, there is a bug insofar as we don't explicitly flag this as
requiring the project to be replaced. This creates an inconsistent plan.