team_id should not be computed
Team IDs are optionally set on a per-resource basis.
However, when omitted, they should not be set as 'computed', because
they can never be anything other than Null/Unknown. Computed makes them
show as unknown, expecting the value to later be set.

This causes a weird churn when creating resources under a personal
Vercel account.

Closes #136