Validate sensitive environment variable config when creating env vars
Turns out the API elides non-sensitive values into being sensitive if
the team has a sensitiveEnvironmentVariablePolicy set on it.

To prevent this from causing provider issues when the `sensitive` field
is explicitly set to `false`, warn at plan time that the resource being
created does not match the team policy.

Closes #183