Consider and use the protection bypass for automation secret from the plan
This fixes #294.

Before this change, the vercel_project resource did not pass on changes
to the protection_bypass_for_automation_secret attribute on to the API.

This commit contains the following changes:
* `client/project_protection_bypass_for_automation_update.go`: Change the anonymous struct member to accurately reflect what is marshalled
* `projectResource.Create()`: always save the user-provided secret in the state, even if it is empty. This is required so that the update logic below doesn't try to reset the secret when it is generated by the vercel API.
* `projectResource.Update()`: correctly handle the different update scenarios


## Limitations

This change stops API-generated secrets to be returned into the state. This
will break existing setups that rely on this undocumented fact.